Greatly improved, just not working anymore.
Added conveyors which will move any Moveables on top of them in a
predefined direction.
Added a function to deal with mouse dragging which initialises new
conveyors in the direction of the dragging.
Added Debug options as global variables, so various debug printouts can
be turned on and off in the code.
Whole thing does not work at the moment due to the GL interpreter in
pyglet getting given a float from somewhere, I have NO IDEA where.